% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{career_compare}
\alias{career_compare}
\title{Compare two career levels.}
\usage{
career_compare(a, b)
}
\arguments{
\item{a, b}{Career level strings.}
}
\value{
Integer: -1 if a < b, 0 if equal, 1 if a > b.
}
\description{
Compare two career levels.
}
\keyword{internal}
